Move interfaces from OpenAi package
