+++ webapp/src/main/java/com/box/l10n/mojito/service/ai/openai/LLMPromptService.java
@@ -10,7 +10,11 @@ import com.box.l10n.mojito.entity.RepositoryAIPrompt;
 import com.box.l10n.mojito.rest.ai.AIException;
 import com.box.l10n.mojito.rest.ai.AIPromptContextMessageCreateRequest;
 import com.box.l10n.mojito.rest.ai.AIPromptCreateRequest;
+import com.box.l10n.mojito.service.ai.AIPromptContextMessageRepository;
+import com.box.l10n.mojito.service.ai.AIPromptRepository;
+import com.box.l10n.mojito.service.ai.AIPromptTypeRepository;
 import com.box.l10n.mojito.service.ai.PromptService;
+import com.box.l10n.mojito.service.ai.RepositoryAIPromptRepository;
 import com.box.l10n.mojito.service.repository.RepositoryRepository;
 import io.micrometer.core.annotation.Timed;
 import jakarta.transaction.Transactional;
