+++ webapp/src/test/java/com/box/l10n/mojito/service/ai/openai/LLMPromptServiceTest.java
@@ -15,6 +15,9 @@ import com.box.l10n.mojito.entity.Repository;
 import com.box.l10n.mojito.entity.RepositoryAIPrompt;
 import com.box.l10n.mojito.rest.ai.AIException;
 import com.box.l10n.mojito.rest.ai.AIPromptCreateRequest;
+import com.box.l10n.mojito.service.ai.AIPromptRepository;
+import com.box.l10n.mojito.service.ai.AIPromptTypeRepository;
+import com.box.l10n.mojito.service.ai.RepositoryAIPromptRepository;
 import com.box.l10n.mojito.service.repository.RepositoryRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
