+++ webapp/src/test/java/com/box/l10n/mojito/service/ai/openai/OpenAILLMServiceTest.java
@@ -12,6 +12,7 @@ import com.box.l10n.mojito.okapi.extractor.AssetExtractorTextUnit;
 import com.box.l10n.mojito.openai.OpenAIClient;
 import com.box.l10n.mojito.rest.ai.AICheckRequest;
 import com.box.l10n.mojito.rest.ai.AICheckResponse;
+import com.box.l10n.mojito.service.ai.AIStringCheckRepository;
 import com.box.l10n.mojito.service.repository.RepositoryRepository;
 import java.util.ArrayList;
 import java.util.List;
